+++ src/system/application/Service/message/browse.php
@@ -11,7 +11,7 @@ trait browse
         $public_key = algo_gen_hash($private_key, SLOPT_PRIVATE_KEY_TO_PUBLIC_KEY);
         $public_key_h = algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_DIRNAME);
 
-        if (!is_private_key_valid($private_key)) {
+        if (!is_hash_valid($private_key)) {
             add_message("error", "Invalid private key.");
             return json_response();
         }
@@ -58,13 +58,14 @@ trait browse
 
         foreach (array_slice($glob, $pagination_data->skip, $pagination_data->take == -1 ? count($glob) : $pagination_data->take) as $message) {
             $message_content = file_get_contents($message);
-            $message_decrypted = json_decode(decrypt_message($message_content, algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_KEY)));
+            $message_decrypted = json_decode(decrypt_message($message_content, algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_SECRET)));
             $data[] = [
                 "id" => $message_decrypted->id,
                 "created_at" => $message_decrypted->manifest->created_at,
                 "is_modified" => $message_decrypted->manifest->is_modified,
                 "from" => $message_decrypted->pair->from,
                 "to" => $message_decrypted->pair->to,
+                "read" => $message_decrypted->read,
                 "message" => [
                     "subject" => substr($message_decrypted->subject, 0, 32),
                     "content" => substr($message_decrypted->content, 0, 32),