+++ src/system/application/Service/message/edit.php
@@ -16,12 +16,12 @@ trait edit
         $public_key = algo_gen_hash($private_key, SLOPT_PRIVATE_KEY_TO_PUBLIC_KEY);
         $public_key_h = algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_DIRNAME);
 
-        if (!is_public_key_valid($public_key)) {
+        if (!is_hash_valid($public_key)) {
             add_message("error", "Invalid public key.");
             return json_response();
         }
 
-        if (!is_private_key_valid($private_key)) {
+        if (!is_hash_valid($private_key)) {
             add_message("error", "Invalid private key.");
             return json_response();
         }
@@ -37,7 +37,7 @@ trait edit
         }
 
         $message_content = file_get_contents($file_path);
-        $message_decrypted = json_decode(decrypt_message($message_content, algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_KEY)));
+        $message_decrypted = json_decode(decrypt_message($message_content, algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_SECRET)));
 
         if (!secure_strcmp($message_decrypted->pair->from, $sender_public_key = algo_gen_hash($private_key, SLOPT_PRIVATE_KEY_TO_PUBLIC_KEY))) {
             add_message("error", "The private key does not match with the sender's private key of the message.");
@@ -60,23 +60,21 @@ trait edit
                 "is_modified" => true,
                 "message_blake3_digest" => blake3($message->content . $message->subject)
             ],
-            "pair" => [
-                "from" => $sender_public_key,
-                "to" => $public_key
-            ]
+            "pair" => $message_decrypted->pair
         ];
-
-        $message_json_data = encrypt_message(json_encode($message_x), algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_KEY));
 
         if (strlen($message->content . $message->subject) >= parse_hsize($size = config("information.message_max_size"))) {
             add_message("error", "Message content cannot be bigger than " . $size . " bytes.");
             return json_response();
         }
 
-        file_put_contents($public_key_d . "/" . algo_gen_hash($id, SLOPT_SKYID_HASH), $message_json_data);
-        $sender_public_key_h = algo_gen_hash($sender_public_key, SLOPT_PUBLIC_KEY_DIRNAME);
-        $message_json_data_for_sender = encrypt_message(json_encode($message_x), algo_gen_hash($sender_public_key, SLOPT_PUBLIC_KEY_KEY));
-        file_put_contents(SENT_PATH . $sender_public_key_h . "/" . algo_gen_hash($id, SLOPT_SKYID_HASH), $message_json_data_for_sender);
+        $storeTo = [$public_key, ...$message_decrypted->pair->to];
+        foreach ($storeTo as $store_public_key) {
+            $basePath = strcmp($store_public_key, $public_key) == 0 ? SENT_PATH : INBOX_PATH;
+            $store_public_key_h = algo_gen_hash($store_public_key, SLOPT_PUBLIC_KEY_DIRNAME);
+            $message_json_data_for_sender = encrypt_message(json_encode($message_x), algo_gen_hash($store_public_key, SLOPT_PUBLIC_KEY_SECRET));
+            file_put_contents($basePath . $store_public_key_h . "/" . algo_gen_hash($id, SLOPT_SKYID_HASH), $message_json_data_for_sender);
+        }
 
         add_message("info", "Message edited using sender's private key");
         return json_response(