+++ src/system/application/Service/message/read.php
@@ -10,15 +10,15 @@ trait read
         $public_key = algo_gen_hash($private_key, SLOPT_PRIVATE_KEY_TO_PUBLIC_KEY);
         $public_key_h = algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_DIRNAME);
 
-        if (!is_private_key_valid($private_key)) {
+        if (!is_hash_valid($private_key)) {
             add_message("error", "Invalid private key.");
             return json_response();
         }
 
         $dir = SENT_PATH;
-    tryagain:
+        tryagain:
         if (!is_dir($public_key_d = $dir . $public_key_h)) {
-            if($dir == SENT_PATH) {
+            if ($dir == SENT_PATH) {
                 $dir = INBOX_PATH;
                 goto tryagain;
             } else {
@@ -26,13 +26,21 @@ trait read
             }
         }
 
-        if (!is_file($file_path = $public_key_d . "/" . algo_gen_hash($id, SLOPT_SKYID_HASH))) {
-            add_message("error", "Message not found");
-            return json_response();
+        $file_path = $public_key_d . "/" . algo_gen_hash($id, SLOPT_SKYID_HASH);
+        if (!is_file($file_path)) {
+            if ($dir == SENT_PATH) {
+                $dir = INBOX_PATH;
+                goto tryagain;
+            } else {
+                add_message("error", "Message not found");
+                return json_response();
+            }
         }
 
+        $key = algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_SECRET);
         $message_content = file_get_contents($file_path);
-        $message_decrypted = json_decode(decrypt_message($message_content, algo_gen_hash($public_key, SLOPT_PUBLIC_KEY_KEY)));
+        $message_decrypted = json_decode(decrypt_message($message_content, $key));
+        $message_decrypted->read = true;
         $data = [
             "id" => $message_decrypted->id,
             "manifest" => $message_decrypted->manifest,
@@ -43,6 +51,8 @@ trait read
                 "content" => $message_decrypted->content,
             ]
         ];
+
+        file_put_contents($file_path, encrypt_message(json_encode($message_decrypted), $key));
 
         return json_response($data);
     }