+++ src/system/application/Service/signing/view.php
@@ -10,14 +10,14 @@ trait view
         $issuer_public_key = algo_gen_hash($private_key, SLOPT_PRIVATE_KEY_TO_PUBLIC_KEY);
         $issuer_public_key_h = algo_gen_hash($issuer_public_key, SLOPT_PUBLIC_KEY_DIRNAME);
 
-        if (!is_private_key_valid($private_key)) {
+        if (!is_hash_valid($private_key)) {
             add_message("error", "Invalid private key.");
             return json_response();
         }
         $path = CONTRACT_FROM_PATH;
-    tryagain:
+        tryagain:
         if (!is_file($file_path = $path . $issuer_public_key_h . '/' . algo_gen_hash($id, SLOPT_SKYID_HASH))) {
-            if($path == CONTRACT_FROM_PATH) {
+            if ($path == CONTRACT_FROM_PATH) {
                 $path = CONTRACT_TO_PATH;
                 goto tryagain;
             } else {
@@ -26,8 +26,8 @@ trait view
             }
         }
 
-        $sign_data = json_decode(decrypt_message(file_get_contents($file_path), algo_gen_hash($issuer_public_key, SLOPT_PUBLIC_KEY_KEY)));
-        
+        $sign_data = json_decode(decrypt_message(file_get_contents($file_path), algo_gen_hash($issuer_public_key, SLOPT_PUBLIC_KEY_SECRET)));
+
         $now = time();
         if ($now + $sign_data->expires < $now && $sign_data->status->sign_status == null) {
             add_message("info", "The receiver did not sign the contract on time.");