+++ src/system/boot/cors.php
@@ -1,0 +1,12 @@
+<?php
+
+header('Access-Control-Allow-Origin: *');
+header('Access-Control-Allow-Methods: *');
+header('Access-Control-Allow-Headers: token, Content-Type');
+header('Access-Control-Max-Age: 1728000');
+
+if ($_SERVER['REQUEST_METHOD'] === 'OPTIONS') {
+    header('Content-Length: 0');
+    header('Content-Type: text/plain');
+    die();
+}